problems final version (everything works 99.9%)

+++ data/submissions.py
@@ -9,7 +9,7 @@ class Submission(SqlAlchemyBase, SerializerMixin):
     __tablename__ = 'submissions'
     id = sqlalchemy.Column(sqlalchemy.Integer, primary_key=True, autoincrement=True)
     verdict = sqlalchemy.Column(sqlalchemy.String)
-    time = sqlalchemy.Column(sqlalchemy.Float)
+    time = sqlalchemy.Column(sqlalchemy.Integer)
     memory = sqlalchemy.Column(sqlalchemy.Integer)
     language = sqlalchemy.Column(sqlalchemy.String)
     send_time = sqlalchemy.Column(sqlalchemy.String, default=datetime.datetime.now().strftime("%d/%b/%Y  %H:%M  UTC+6"))
